requirements.txt
discord.py
rethinkdb